+++ src/features/migration/MigrationEntries.ts
@@ -10,16 +10,35 @@ import type { MigratableEntry, TMigrationEntry } from '@/features/migration/Migr
 import { MigrationEntryStatus } from '@/features/migration/Migration.types.ts';
 
 export class MigrationEntries {
+    public static isAbortable(entry: TMigrationEntry): boolean {
+        return (
+            MigrationEntries.isSearching(entry) ||
+            MigrationEntries.hasStatus(entry, MigrationEntryStatus.MIGRATION_PENDING)
+        );
+    }
+
     public static isSearching(entry: TMigrationEntry): boolean {
-        return [MigrationEntryStatus.PENDING, MigrationEntryStatus.SEARCHING].includes(entry.status);
+        return MigrationEntries.hasStatus(entry, MigrationEntryStatus.SEARCH_PENDING, MigrationEntryStatus.SEARCHING);
+    }
+
+    public static isMigrating(entry: TMigrationEntry): boolean {
+        return MigrationEntries.hasStatus(
+            entry,
+            MigrationEntryStatus.MIGRATION_PENDING,
+            MigrationEntryStatus.MIGRATING,
+        );
     }
 
     public static getExcluded(entries: TMigrationEntry[]): TMigrationEntry[] {
         return entries.filter((entry) => entry.isExcluded);
     }
 
+    public static hasStatus(entry: TMigrationEntry, ...statuses: MigrationEntryStatus[]): boolean {
+        return statuses.includes(entry.status);
+    }
+
     public static getHaveStatus(entries: TMigrationEntry[], ...statuses: MigrationEntryStatus[]): TMigrationEntry[] {
-        return entries.filter((entry) => statuses.includes(entry.status));
+        return entries.filter((entry) => MigrationEntries.hasStatus(entry, ...statuses));
     }
 
     public static getHaveStatusSorted(
@@ -52,7 +71,12 @@ export class MigrationEntries {
     public static getMigratable(entries: TMigrationEntry[]): MigratableEntry[] {
         return Object.values(entries).filter(
             (entry): entry is MigratableEntry =>
-                [MigrationEntryStatus.SEARCH_COMPLETE, MigrationEntryStatus.MIGRATING].includes(entry.status) &&
+                MigrationEntries.hasStatus(
+                    entry,
+                    MigrationEntryStatus.SEARCH_COMPLETE,
+                    MigrationEntryStatus.MIGRATION_PENDING,
+                    MigrationEntryStatus.MIGRATING,
+                ) &&
                 !entry.isExcluded &&
                 entry.selectedMatchMangaId != null &&
                 entry.selectedMatchSourceId != null,