+++ src/features/migration/components/migration-entry/MigrationEntrySearchExcludeActions.tsx
@@ -15,8 +15,6 @@ import IconButton from '@mui/material/IconButton';
 import { useLingui } from '@lingui/react/macro';
 import Stack from '@mui/material/Stack';
 import SearchIcon from '@mui/icons-material/Search';
-import CloseIcon from '@mui/icons-material/Close';
-import AddIcon from '@mui/icons-material/Add';
 import Button from '@mui/material/Button';
 import { plural } from '@lingui/core/macro';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
@@ -24,6 +22,7 @@ import ExpandLessIcon from '@mui/icons-material/ExpandLess';
 import ButtonGroup from '@mui/material/ButtonGroup';
 import RemoveCircleOutlineOutlinedIcon from '@mui/icons-material/RemoveCircleOutlineOutlined';
 import AddCircleOutlineOutlinedIcon from '@mui/icons-material/AddCircleOutlineOutlined';
+import CancelOutlinedIcon from '@mui/icons-material/CancelOutlined';
 
 export const MigrationEntrySearchExcludeActions = ({
     hasSelectedMatch,
@@ -33,6 +32,8 @@ export const MigrationEntrySearchExcludeActions = ({
     isExcluded,
     mangaId,
     mangaTitle,
+    isAbortable,
+    isMigrating,
 }: {
     hasSelectedMatch: boolean;
     otherResultsCount: number;
@@ -41,6 +42,8 @@ export const MigrationEntrySearchExcludeActions = ({
     isExcluded: boolean;
     mangaId: MangaIdInfo['id'];
     mangaTitle: string;
+    isAbortable: boolean;
+    isMigrating: boolean;
 }) => {
     const { t } = useLingui();
     const isTabletWidth = MediaQuery.useIsTabletWidth();
@@ -48,7 +51,7 @@ export const MigrationEntrySearchExcludeActions = ({
     if (isTabletWidth) {
         return (
             <ButtonGroup variant="contained">
-                {!!otherResultsCount && (
+                {!isMigrating && !!otherResultsCount && (
                     <Button
                         sx={{ flexGrow: 1 }}
                         startIcon={isExpanded ? <ExpandLessIcon /> : <ExpandMoreIcon />}
@@ -60,7 +63,7 @@ export const MigrationEntrySearchExcludeActions = ({
                         })}
                     </Button>
                 )}
-                {!isExpanded && (
+                {!isMigrating && !isExpanded && (
                     <CustomTooltip title={t`Manual search`}>
                         <CustomButtonIcon
                             sx={{
@@ -74,7 +77,7 @@ export const MigrationEntrySearchExcludeActions = ({
                         </CustomButtonIcon>
                     </CustomTooltip>
                 )}
-                {hasSelectedMatch && (
+                {!isMigrating && hasSelectedMatch && (
                     <CustomTooltip title={isExcluded ? t`Include` : t`Exclude`}>
                         <CustomButtonIcon
                             sx={{
@@ -90,32 +93,53 @@ export const MigrationEntrySearchExcludeActions = ({
                         </CustomButtonIcon>
                     </CustomTooltip>
                 )}
+                {isAbortable && (
+                    <CustomTooltip title={t`Abort`}>
+                        <CustomButtonIcon
+                            sx={{
+                                flexGrow: Number(!otherResultsCount),
+                            }}
+                            onClick={() => MigrationManager.abortEntry(mangaId)}
+                        >
+                            <CancelOutlinedIcon />
+                        </CustomButtonIcon>
+                    </CustomTooltip>
+                )}
             </ButtonGroup>
         );
     }
 
     return (
         <Stack sx={{ gap: 1, justifyContent: 'center' }}>
-            {hasSelectedMatch && (
+            {isAbortable && (
+                <CustomTooltip title={t`Abort`} placement="auto">
+                    <IconButton onClick={() => MigrationManager.abortEntry(mangaId)}>
+                        <CancelOutlinedIcon />
+                    </IconButton>
+                </CustomTooltip>
+            )}
+            {!isMigrating && hasSelectedMatch && (
                 <CustomTooltip title={isExcluded ? t`Include` : t`Exclude`} placement="auto">
                     <IconButton
                         onClick={() =>
                             isExcluded ? MigrationManager.includeManga(mangaId) : MigrationManager.excludeManga(mangaId)
                         }
                     >
-                        {isExcluded ? <AddIcon /> : <CloseIcon />}
+                        {isExcluded ? <AddCircleOutlineOutlinedIcon /> : <RemoveCircleOutlineOutlinedIcon />}
                     </IconButton>
                 </CustomTooltip>
             )}
-            <CustomTooltip title={t`Manual search`} placement="auto">
-                <IconButton
-                    onClick={() => {
-                        MigrationManager.openManualSearch(mangaId, mangaTitle);
-                    }}
-                >
-                    <SearchIcon />
-                </IconButton>
-            </CustomTooltip>
+            {!isMigrating && (
+                <CustomTooltip title={t`Manual search`} placement="auto">
+                    <IconButton
+                        onClick={() => {
+                            MigrationManager.openManualSearch(mangaId, mangaTitle);
+                        }}
+                    >
+                        <SearchIcon />
+                    </IconButton>
+                </CustomTooltip>
+            )}
         </Stack>
     );
 };